Reverting the bishop-pair evaluation term in preperation of the release of v2.2
Revert "v2.1.11d does only update the bishop pair evaluation term outside qsearch but now is a little more complex. It still has the fixed bishop-pair bonus of 25 cp but also applies a bonus/malus with only one bishop present. In that case the score modifier is based on whether the opponent pieces are predominantly standing on same-colored squares or opposite-colored squares as the bishop."

This reverts commit 7dadabe506f0e958bdbe2a2f1c74c1acc7e5ecc1.

Revert "Experimenting with BishopPair bonus. This updates the bonus even in QSearch and was used to build 2.1.11c"

This reverts commit 53acb6de8d705dbe3651381e60e22d5e1e23b6d1.