Futility pruning has been extended over all plies. The margins (multiplied with remaining moves) can be configured using UCI options. The default values FutilityMargin = 90, LateFutilityMargin = 50 have been found using https://github.com/kiudee/chess-tuning-tools
The result is a little bitter sweet. In self play the new version gained 73 Elo which is hard to belief: Score of Leorik-2.3.4a vs Leorik-2.3.3c: 4050 - 1958 - 3992  [0.605] 10000
There's an Elo gain in the gauntlet on longer time controls of 20-30 Elo.
On the WAC test the new version solves 10 less positions. Basically it misses out on all the interesting sacrifices.

Needs further investigation but still worth to be preserved in a commit.