Added Temperature & Seed property to SearchOptions. Each root move now gets assigned a random bonus from the range [0..Temperature] which will make the engine play a greater variety of lines but decreases playing strength.
Created additional labeled positions with Temperature=50, removed the 9M cap and tuned on 17M positions instead. Validation set is no longer constant but changes 10% per iteration which may or may not be a good idea. But all in all the new weights are the largest improvement over 2.5.4a so far and deserve to be preserved: 3185 - 2666 - 4149  [0.526] 10000