256HL-3B5083B8.nnue is stronger than it looked. It just didn't work well with the old default NullMoveCutoff value. Removing the early out based on NullMoveCutoff is worth +60 Elo. [0.854] vs  [0.799] against Leorik 2.5.6
But this early out was gaining and actually still gains for 128HL-57716AF0.nnue (+20) Elo so the feature is not to blame but the threshold doesn't fit anymore.

I hacked IterativeSearch to keep track of how good that prediction of a beta cutoff is. 128HL-57716AF0 achieved a FALSE_SKIP/CORRECT ratio of ~10%. To get similar stats with the new 256 HL net you have to double the threshold. This commit actually does something based on beta. But anyway... with adjusted threshold 256HL-3B5083B8 the feature is very strong:

[0.889] against Leorik 2.5.6 aka 360 Elo.

If this hold for play against other engines it should put me in the range of Zahak 9.0 which is 3300 CCRL.
And a quick test confirms it: Score of Leorik-256HL-B vs zahak900: 393 - 423 - 659  [0.490] 1475