When doing null move pruning I skip the null move evaluation for "obvious" situation where static eval is above beta + margin. But such a margin depends on the evaluation and needs to be retuned when a new network is used.
This commit is about computing the margin dynamically based on live statistics gathered while searching the current position. This decouples the search from evaluation and gets rid of a "magic constant" that can otherwise only be found by trial and error and where the best value is likely not really constant to begin with.

The first and simplest implementation already seems promising!

Score of Leorik-3.0.5 vs Leorik-3.0.4: 1619 - 1501 - 5581  [0.507] 8701
...      Leorik-3.0.5 playing White: 1000 - 547 - 2803  [0.552] 4350
...      Leorik-3.0.5 playing Black: 619 - 954 - 2778  [0.462] 4351
...      White vs Black: 1954 - 1166 - 5581  [0.545] 8701
Elo difference: 4.7 +/- 4.4, LOS: 98.3 %, DrawRatio: 64.1 %