Changed MoveGen to support Chess960 castling so that perft() results on fischer.epd are all correct.
Castling is different in Chess960 and the white king move when castling short is not always e1g1.
A king move could both be the castling king move or just a normal king move.
This is why castling moves are sent in the form king "takes" his own rook.
Example: e1h1 for the white short castle move in the normal chess start position.

This is still WIP: the implementation has room for optimization and the engine should support both Classical and FRC at the same time.